# docs/swagger.yaml
basePath: /
definitions:
  domain.CreateOrderItemRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  domain.CreateOrderRequest:
    properties:
      customer_id:
        type: integer
      items:
        items:
          $ref: '#/definitions/domain.CreateOrderItemRequest'
        minItems: 1
        type: array
    required:
    - customer_id
    - items
    type: object
  domain.Customer:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    type: object
  domain.Product:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
      updated_at:
        type: string
    required:
    - name
    - price
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  domain.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  domain.UserLoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.UserRegistrationRequest:
    properties:
      password:
        minLength: 6
        type: string
      role:
        type: string
      username:
        minLength: 4
        type: string
    required:
    - password
    - username
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  usecase.DashboardStats:
    properties:
      total_customers:
        type: integer
      total_orders:
        type: integer
      total_revenue:
        type: number
    type: object
host: localhost:8080
info:
  contact:
    email: support@simpleorder.local
    name: API Support
  description: This is a sample production-ready backend project for SimpleOrder.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SimpleOrder API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserLoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login a user
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UserRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - Auth
  /api/v1/customers:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name or email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Customers
    post:
      consumes:
      - application/json
      parameters:
      - description: Customer Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Customer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new customer
      tags:
      - Customers
  /api/v1/dashboard/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/usecase.DashboardStats'
              type: object
      security:
      - BearerAuth: []
      summary: Get Dashboard Stats
      tags:
      - Dashboard
  /api/v1/orders:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Order Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
  /api/v1/orders/{id}/invoice:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Generate Order Invoice PDF
      tags:
      - Orders
  /api/v1/products:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"